Merging 1.6.8 into master

+++ lib/agents/ec2_agent.py
@@ -1,4 +1,9 @@
-from agents.base_agent import BaseAgent, AgentConfigurationException, AgentRuntimeException
+"""
+Helper library for EC2 interaction
+"""
+from agents.base_agent import AgentRuntimeException
+from agents.base_agent import BaseAgent 
+from agents.base_agent import AgentConfigurationException
 import boto
 from boto.exception import EC2ResponseError
 import datetime
@@ -6,9 +11,10 @@ import os
 import time
 from appscale_logger import AppScaleLogger
 from local_state import LocalState
+# pylint: disable-msg=W0511
+#    don't bother about todo's
 
-__author__ = 'hiranya'
-__email__ = 'hiranya@appscale.com'
+
 
 class EC2Agent(BaseAgent):
   """
@@ -75,7 +81,7 @@ class EC2Agent(BaseAgent):
     the parameters dictionary. More specifically, this method expects to
     find a 'keyname' parameter and a 'group' parameter in the parameters
     dictionary. Using these provided values, this method will create a new
-    EC2 key-pair and a security group. Security group will be granted permissions
+    EC2 key-pair and a security group. Security group will be granted permission
     to access any port on the instantiated VMs. (Also see documentation for the
     BaseAgent class)
 
@@ -125,20 +131,29 @@ class EC2Agent(BaseAgent):
       args: A Namespace containing the arguments that the user has
         invoked an AppScale Tool with.
     """
+    # need to convert this to a dict if it is not already
+    if not isinstance(args, dict):
+      args = vars(args)
+
     params = {
       self.PARAM_CREDENTIALS : {},
-      self.PARAM_GROUP : args.group,
-      self.PARAM_IMAGE_ID : args.machine,
-      self.PARAM_INSTANCE_TYPE : args.instance_type,
-      self.PARAM_KEYNAME : args.keyname,
-      'IS_VERBOSE' : args.verbose
+      self.PARAM_GROUP : args['group'],
+      self.PARAM_IMAGE_ID : args['machine'],
+      self.PARAM_INSTANCE_TYPE : args['instance_type'],
+      self.PARAM_KEYNAME : args['keyname'],
     }
+    if 'verbose' in args:
+      params['IS_VERBOSE'] = args['verbose']
+    else:
+      params['IS_VERBOSE'] = False
 
     for credential in self.REQUIRED_CREDENTIALS:
-      if os.environ[credential] and os.environ[credential] != '':
+      if credential in os.environ and os.environ[credential] != '':
         params[self.PARAM_CREDENTIALS][credential] = os.environ[credential]
       else:
-        raise AgentConfigurationException("no " + credential)
+        raise AgentConfigurationException("Couldn't find {0} in your " \
+          "environment. Please set it and run AppScale again."
+          .format(credential))
 
     return params
 
@@ -228,8 +243,9 @@ class EC2Agent(BaseAgent):
 
     Args:
       count               No. of VMs to spawned
-      parameters          A dictionary of parameters. This must contain 'keyname',
-                          'group', 'image_id' and 'instance_type' parameters.
+      parameters          A dictionary of parameters. This must contain 
+                          'keyname', 'group', 'image_id' and 'instance_type'
+                          parameters.
       security_configured Uses this boolean value as an heuristic to
                           detect brand new AppScale deployments.
 
@@ -242,8 +258,13 @@ class EC2Agent(BaseAgent):
     group = parameters[self.PARAM_GROUP]
     spot = False
 
-    AppScaleLogger.log('[{0}] [{1}] [{2}] [{3}] [ec2] [{4}] [{5}]'.format(count,
-      image_id, instance_type, keyname, group, spot))
+    AppScaleLogger.log("Starting {0} machines with machine id {1}, with " \
+      "instance type {2}, keyname {3}, in security group {4}".format(count,
+      image_id, instance_type, keyname, group))
+    if spot:
+      AppScaleLogger.log("Using spot instances")
+    else:
+      AppScaleLogger.log("Using on-demand instances")
 
     start_time = datetime.datetime.now()
     active_public_ips = []
@@ -285,7 +306,6 @@ class EC2Agent(BaseAgent):
 
       while now < end_time:
         time_left = (end_time - now).seconds
-        AppScaleLogger.log('[{0}] {1} seconds left...'.format(now, time_left))
         instance_info = self.describe_instances(parameters)
         public_ips = instance_info[0]
         private_ips = instance_info[1]
@@ -306,26 +326,26 @@ class EC2Agent(BaseAgent):
         for index in range(0, len(public_ips)):
           if public_ips[index] == '0.0.0.0':
             instance_to_term = instance_ids[index]
-            AppScaleLogger.log('Instance {0} failed to get a public IP address and' \
-                      ' is being terminated'.format(instance_to_term))
+            AppScaleLogger.log('Instance {0} failed to get a public IP address'\
+                    'and is being terminated'.format(instance_to_term))
             conn.terminate_instances([instance_to_term])
 
       end_time = datetime.datetime.now()
       total_time = end_time - start_time
       if spot:
-        AppScaleLogger.log('TIMING: It took {0} seconds to spawn {1} spot ' \
-                  'instances'.format(total_time.seconds, count))
+        AppScaleLogger.log("Started {0} spot instances in {1} seconds" \
+          .format(count, total_time.seconds))
       else:
-        AppScaleLogger.log('TIMING: It took {0} seconds to spawn {1} ' \
-                  'regular instances'.format(total_time.seconds, count))
+        AppScaleLogger.log("Started {0} on-demand instances in {1} seconds" \
+          .format(count, total_time.seconds))
       return instance_ids, public_ips, private_ips
     except EC2ResponseError as exception:
       self.handle_failure('EC2 response error while starting VMs: ' +
                           exception.error_message)
 
-  def terminate_instances(self, parameters):
+  def stop_instances(self, parameters):
     """
-    Stop one of more EC2 instances using. The input instance IDs are
+    Stop one of more EC2 instances. The input instance IDs are
     fetched from the 'instance_ids' parameters in the input map. (Also
     see documentation for the BaseAgent class)
 
@@ -334,9 +354,92 @@ class EC2Agent(BaseAgent):
     """
     instance_ids = parameters[self.PARAM_INSTANCE_IDS]
     conn = self.open_connection(parameters)
-    terminated_instances = conn.terminate_instances(instance_ids)
-    for instance in terminated_instances:
-      AppScaleLogger.log('Instance {0} was terminated'.format(instance.id))
+    conn.stop_instances(instance_ids)
+    AppScaleLogger.log('Stopping instances: '+' '.join(instance_ids))
+    if not self.wait_for_status_change(parameters, conn, 'stopped',\
+           max_wait_time=120):
+      AppScaleLogger.log("re-stopping instances: "+' '.join(instance_ids))
+      conn.stop_instances(instance_ids)
+      if not self.wait_for_status_change(parameters, conn, 'stopped',\
+            max_wait_time=120):
+        self.handle_failure("ERROR: could not stop instances: "+\
+            ' '.join(instance_ids))
+
+
+  def terminate_instances(self, parameters):
+    """
+    Terminate one of more EC2 instances. The input instance IDs are
+    fetched from the 'instance_ids' parameters in the input map. (Also
+    see documentation for the BaseAgent class)
+
+    Args:
+      parameters:  A dictionary of parameters
+    """
+    instance_ids = parameters[self.PARAM_INSTANCE_IDS]
+    conn = self.open_connection(parameters)
+    conn.terminate_instances(instance_ids)
+    AppScaleLogger.log('Terminating instances: '+' '.join(instance_ids))
+    if not self.wait_for_status_change(parameters, conn, 'terminated',\
+            max_wait_time=120):
+      AppScaleLogger.log("re-terminating instances: "+' '.join(instance_ids))
+      conn.terminate_instances(instance_ids)
+      if not self.wait_for_status_change(parameters, conn, 'terminated',\
+                max_wait_time=120):
+        self.handle_failure("ERROR: could not terminate instances: "+\
+            ' '.join(instance_ids))
+
+
+  def wait_for_status_change(self, parameters, conn, state_requested, \
+                              max_wait_time=60,poll_interval=10):
+    """
+    After we have sent a signal to the cloud infrastructure to change the state
+      of the instances (unsually from runnning to either stoppped or 
+      terminated), wait for the status to change.  If all the instances change
+      successfully, return true, if not return false
+    Args:
+      parameters: A dictionary of parameters
+      conn:       A connection object returned from self.open_connection()
+      state_requrested: string of the requested final state of the instances
+      max_wait_time: int of maximum amount of time (in seconds)  to wait for the
+                        state change
+      poll_interval: int of the number of seconds to wait between checking of
+                        the state
+    """
+    time_start = time.time()
+    instance_ids = parameters[self.PARAM_INSTANCE_IDS]
+    instances_in_state = {}
+    while True:
+      time.sleep(poll_interval)
+      reservations = conn.get_all_instances(instance_ids)
+      instances = [i for r in reservations for i in r.instances]
+      for i in instances:
+        # instance i.id reports status = i.state
+        if i.state == state_requested and \
+           i.key_name == parameters[self.PARAM_KEYNAME]:
+          if i.id not in instances_in_state.keys():
+            instances_in_state[i.id] = 1 #mark instance done
+      if len(instances_in_state.keys()) >= len(instance_ids):
+        return True
+      if time.time()-time_start > max_wait_time:
+        return False
+
+
+
+
+  def create_image(self, instance_id, name, parameters):
+    """
+    Take the instance id and name, and creates a cloud image
+    
+    Args:
+      instance_id: id of the (stopped) instance to create an image of
+      name: human readable name for the image
+    Return:
+      str containing the id of the new image
+     """
+    conn = self.open_connection(parameters)
+    id_str = conn.create_image(instance_id, name)
+    return id_str
+
 
   def does_image_exist(self, parameters):
     """
@@ -383,8 +486,9 @@ class EC2Agent(BaseAgent):
   def get_optimal_spot_price(self, conn, instance_type):
     """
     Returns the spot price for an EC2 instance of the specified instance type.
-    The returned value is computed by averaging all the spot price history values
-    returned by the back-end EC2 APIs and incrementing the average by extra 20%.
+    The returned value is computed by averaging all the spot price history
+    values returned by the back-end EC2 APIs and incrementing the average by
+    extra 20%.
 
     Args:
       instance_type An EC2 instance type
@@ -394,13 +498,13 @@ class EC2Agent(BaseAgent):
     """
     history = conn.get_spot_price_history(product_description='Linux/UNIX',
       instance_type=instance_type)
-    sum = 0.0
+    var_sum = 0.0
     for entry in history:
-      sum += entry.price
-    average = sum / len(history)
+      var_sum += entry.price
+    average = var_sum / len(history)
     plus_twenty = average * 1.20
-    AppScaleLogger.log('The average spot instance price for a {0} machine is {1}, '\
-              'and 20% more is {2}'.format(instance_type, average, plus_twenty))
+    AppScaleLogger.log('The average spot instance price for a {0} machine is'\
+        ' {1}, and 20% more is {2}'.format(instance_type, average, plus_twenty))
     return plus_twenty
 
   def open_connection(self, parameters):
